restructure navbar menu
