+++ src/components/CategoryDropdown.js
@@ -1,0 +1,64 @@
+import React from "react";
+import Popper from "popper.js";
+import CategoryDropdownMenu from "./CategoryDropdownMenu";
+function CategoryDropdown(props) {
+  const { menuOpen } = props;
+  const [catDropDown, setCatDropDown] = React.useState(false);
+  const catBtnDropdownRef = React.createRef();
+  const catDropdownRef = React.createRef();
+  const openCatDropdown = () => {
+    new Popper(catBtnDropdownRef.current, catDropdownRef.current, {
+      placement: "auto",
+    });
+    setCatDropDown(true);
+  };
+  const closeCatDropdown = () => {
+    setCatDropDown(false);
+  };
+  if (menuOpen) {
+    return (
+      <>
+        <span
+          onClick={() => {
+            catDropDown ? closeCatDropdown() : openCatDropdown();
+          }}
+          ref={catBtnDropdownRef}
+          type="submit"
+          className="text-black hover:bg-blue-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
+        >
+          Category
+        </span>
+        <CategoryDropdownMenu
+          catDropdownRef={catDropdownRef}
+          catDropDown={catDropDown}
+        />
+      </>
+    );
+  }
+  return (
+    <div
+      onMouseEnter={() => {
+        catDropDown ? closeCatDropdown() : openCatDropdown();
+      }}
+      onMouseLeave={() => {
+        if (catDropDown) closeCatDropdown();
+      }}
+    >
+      <button
+        onClick={() => {
+          catDropDown ? closeCatDropdown() : openCatDropdown();
+        }}
+        ref={catBtnDropdownRef}
+        className="text-black hover:bg-blue-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium"
+      >
+        Category
+      </button>
+      <CategoryDropdownMenu
+        catDropdownRef={catDropdownRef}
+        catDropDown={catDropDown}
+      />
+    </div>
+  );
+}
+
+export default CategoryDropdown;
